+++ tests/unit/test_portia.py
@@ -25,7 +25,6 @@ from portia.clarification import (
 )
 from portia.clarification_handler import ClarificationHandler
 from portia.config import (
-    FEATURE_FLAG_AGENT_MEMORY_ENABLED,
     Config,
     GenerativeModelsConfig,
     StorageClass,
@@ -46,7 +45,6 @@ from portia.introspection_agents.introspection_agent import (
     PreStepIntrospection,
     PreStepIntrospectionOutcome,
 )
-from portia.model import GenerativeModel
 from portia.open_source_tools.llm_tool import LLMTool
 from portia.open_source_tools.registry import example_tool_registry, open_source_tool_registry
 from portia.plan import (
@@ -65,7 +63,6 @@ from portia.planning_agents.base_planning_agent import StepsOrError
 from portia.portia import ExecutionHooks, Portia
 from portia.prefixed_uuid import ClarificationUUID
 from portia.storage import StorageError
-from portia.telemetry.telemetry_service import BaseProductTelemetry
 from portia.telemetry.views import PortiaFunctionCallTelemetryEvent
 from portia.tool import (
     PortiaRemoteTool,
@@ -90,57 +87,6 @@ if TYPE_CHECKING:
     from pytest_httpx import HTTPXMock
 
     from portia.common import Serializable
-
-
-@pytest.fixture
-def telemetry() -> MagicMock:
-    """Fixture to create ProductTelemetry mock."""
-    return MagicMock(spec=BaseProductTelemetry)
-
-
-@pytest.fixture
-def planning_model() -> MagicMock:
-    """Fixture to create a mock planning model."""
-    return MagicMock(spec=GenerativeModel)
-
-
-@pytest.fixture
-def default_model() -> MagicMock:
-    """Fixture to create a mock default model."""
-    return MagicMock(spec=GenerativeModel)
-
-
-@pytest.fixture
-def portia(planning_model: MagicMock, default_model: MagicMock, telemetry: MagicMock) -> Portia:
-    """Fixture to create a Portia instance for testing."""
-    config = get_test_config(
-        models=GenerativeModelsConfig(
-            planning_model=planning_model,
-            default_model=default_model,
-        ),
-    )
-    tool_registry = ToolRegistry([AdditionTool(), ClarificationTool()])
-    return Portia(config=config, tools=tool_registry, telemetry=telemetry)
-
-
-@pytest.fixture
-def portia_with_agent_memory(
-    planning_model: MagicMock, default_model: MagicMock, telemetry: MagicMock
-) -> Portia:
-    """Fixture to create a Portia instance for testing."""
-    config = get_test_config(
-        # Set a small threshold value so all outputs are stored in agent memory
-        feature_flags={FEATURE_FLAG_AGENT_MEMORY_ENABLED: True},
-        large_output_threshold_tokens=3,
-        models=GenerativeModelsConfig(
-            planning_model=planning_model,
-            default_model=default_model,
-        ),
-        portia_api_endpoint="https://api.portialabs.ai",
-        portia_api_key="test-api-key",
-    )
-    tool_registry = ToolRegistry([AdditionTool(), ClarificationTool()])
-    return Portia(config=config, tools=tool_registry, telemetry=telemetry)
 
 
 def test_portia_local_default_config_with_api_keys() -> None:
@@ -3093,279 +3039,3 @@ def test_portia_resolve_example_plans_with_plan_id_strings(portia: Portia) -> No
     assert len(resolved_plans) == 2
     assert resolved_plans[0].id == example_plan_1.id
     assert resolved_plans[1].id == example_plan_2.id
-
-
-@pytest.mark.asyncio
-async def test_portia_agenerate_plan(
-    portia: Portia, planning_model: MagicMock, telemetry: MagicMock
-) -> None:
-    """Test async planning a query."""
-    query = "example query"
-
-    planning_model.aget_structured_response.return_value = StepsOrError(steps=[], error=None)
-    plan = await portia.aplan(query)
-
-    telemetry.capture.assert_called_once_with(
-        PortiaFunctionCallTelemetryEvent(
-            function_name="portia_aplan",
-            function_call_details={
-                "tools": None,
-                "example_plans_provided": False,
-                "end_user_provided": False,
-                "plan_inputs_provided": False,
-            },
-            name="portia_function_call",
-        )
-    )
-
-    assert plan.plan_context.query == query
-
-
-@pytest.mark.asyncio
-async def test_portia_agenerate_plan_error(
-    portia: Portia, planning_model: MagicMock, telemetry: MagicMock
-) -> None:
-    """Test async planning a query that returns an error."""
-    query = "example query"
-
-    planning_model.aget_structured_response.return_value = StepsOrError(
-        steps=[],
-        error="could not plan",
-    )
-    with pytest.raises(PlanError):
-        await portia.aplan(query)
-
-    # Check that the telemetry event was captured despite the error.
-    telemetry.capture.assert_called_once_with(
-        PortiaFunctionCallTelemetryEvent(
-            function_name="portia_aplan",
-            function_call_details={
-                "tools": None,
-                "example_plans_provided": False,
-                "end_user_provided": False,
-                "plan_inputs_provided": False,
-            },
-            name="portia_function_call",
-        )
-    )
-
-
-@pytest.mark.asyncio
-async def test_portia_agenerate_plan_with_tools(
-    portia: Portia, planning_model: MagicMock, telemetry: MagicMock
-) -> None:
-    """Test async planning a query with tools."""
-    query = "example query"
-
-    planning_model.aget_structured_response.return_value = StepsOrError(steps=[], error=None)
-    plan = await portia.aplan(query, tools=["add_tool"])
-
-    telemetry.capture.assert_called_once_with(
-        PortiaFunctionCallTelemetryEvent(
-            function_name="portia_aplan",
-            function_call_details={
-                "tools": "add_tool",
-                "example_plans_provided": False,
-                "end_user_provided": False,
-                "plan_inputs_provided": False,
-            },
-            name="portia_function_call",
-        )
-    )
-
-    assert plan.plan_context.query == query
-    assert plan.plan_context.tool_ids == ["add_tool"]
-
-
-@pytest.mark.asyncio
-async def test_portia_aplan_with_use_cached_plan_success(portia: Portia) -> None:
-    """Test async planning with use_cached_plan=True when cached plan exists."""
-    query = "example query"
-
-    # Create a cached plan
-    cached_plan = Plan(
-        plan_context=PlanContext(query=query, tool_ids=["add_tool"]),
-        steps=[],
-    )
-    portia.storage.save_plan(cached_plan)
-
-    # Mock the storage.get_plan_by_query to return the cached plan
-    with mock.patch.object(
-        portia.storage, "get_plan_by_query", return_value=cached_plan
-    ) as mock_get_cached:
-        plan = await portia.aplan(query, use_cached_plan=True)
-
-        # Verify get_plan_by_query was called
-        mock_get_cached.assert_called_once_with(query)
-
-        # Verify the cached plan was returned
-        assert plan.id == cached_plan.id
-        assert plan.plan_context.query == query
-
-
-@pytest.mark.asyncio
-async def test_portia_aplan_with_use_cached_plan_not_found(
-    portia: Portia, planning_model: MagicMock
-) -> None:
-    """Test async planning with use_cached_plan=True when no cached plan exists."""
-    query = "example query"
-
-    # Mock the storage.get_plan_by_query to raise StorageError
-    with mock.patch.object(
-        portia.storage, "get_plan_by_query", side_effect=StorageError("No plan found for query")
-    ) as mock_get_cached:
-        # Mock the planning model to return a successful plan
-        planning_model.aget_structured_response.return_value = StepsOrError(steps=[], error=None)
-
-        plan = await portia.aplan(query, use_cached_plan=True)
-
-        # Verify get_plan_by_query was called
-        mock_get_cached.assert_called_once_with(query)
-
-        # Verify a new plan was generated (not the cached one)
-        assert plan.plan_context.query == query
-        assert plan.id != "plan-00000000-0000-0000-0000-000000000000"  # Not a default UUID
-
-
-@pytest.mark.asyncio
-async def test_portia_aplan_with_use_cached_plan_false(
-    portia: Portia, planning_model: MagicMock
-) -> None:
-    """Test async planning with use_cached_plan=False (default behavior)."""
-    query = "example query"
-
-    # Create a cached plan
-    cached_plan = Plan(
-        plan_context=PlanContext(query=query, tool_ids=["add_tool"]),
-        steps=[],
-    )
-    portia.storage.save_plan(cached_plan)
-
-    # Mock the planning model to return a successful plan
-    planning_model.aget_structured_response.return_value = StepsOrError(steps=[], error=None)
-
-    # Mock the storage.get_plan_by_query to ensure it's not called
-    with mock.patch.object(portia.storage, "get_plan_by_query") as mock_get_cached:
-        plan = await portia.aplan(query, use_cached_plan=False)
-
-        # Verify get_plan_by_query was NOT called
-        mock_get_cached.assert_not_called()
-
-        # Verify a new plan was generated
-        assert plan.plan_context.query == query
-        assert plan.id != cached_plan.id  # Should be a different plan
-
-
-@pytest.mark.asyncio
-async def test_portia_aplan_with_use_cached_plan_and_tools(portia: Portia) -> None:
-    """Test async planning with use_cached_plan=True and tools when cached plan exists."""
-    query = "example query"
-
-    # Create a cached plan with tools
-    cached_plan = Plan(
-        plan_context=PlanContext(query=query, tool_ids=["add_tool", "subtract_tool"]),
-        steps=[],
-    )
-    portia.storage.save_plan(cached_plan)
-
-    # Mock the storage.get_plan_by_query to return the cached plan
-    with mock.patch.object(
-        portia.storage, "get_plan_by_query", return_value=cached_plan
-    ) as mock_get_cached:
-        plan = await portia.aplan(query, tools=["add_tool"], use_cached_plan=True)
-
-        # Verify get_plan_by_query was called
-        mock_get_cached.assert_called_once_with(query)
-
-        # Verify the cached plan was returned (tools parameter should be ignored when using
-        # cached plan)
-        assert plan.id == cached_plan.id
-        assert plan.plan_context.query == query
-        assert plan.plan_context.tool_ids == ["add_tool", "subtract_tool"]  # Original cached tools
-
-
-@pytest.mark.asyncio
-async def test_portia_aplan_with_use_cached_plan_storage_error_logging(
-    portia: Portia, planning_model: MagicMock
-) -> None:
-    """Test async planning with use_cached_plan=True when storage error occurs."""
-    query = "example query"
-
-    # Mock the storage.get_plan_by_query to raise StorageError
-    with mock.patch.object(
-        portia.storage, "get_plan_by_query", side_effect=StorageError("Storage error")
-    ) as mock_get_cached:
-        # Mock the planning model to return a successful plan
-        planning_model.aget_structured_response.return_value = StepsOrError(steps=[], error=None)
-
-        plan = await portia.aplan(query, use_cached_plan=True)
-
-        # Verify get_plan_by_query was called
-        mock_get_cached.assert_called_once_with(query)
-
-        # Verify a new plan was generated despite the storage error
-        assert plan.plan_context.query == query
-        assert plan.id != "plan-00000000-0000-0000-0000-000000000000"  # Not a default UUID
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize(
-    "plan_inputs",
-    [
-        [
-            PlanInput(name="$num_a", description="Number A"),
-            PlanInput(name="$num_b", description="Number B"),
-        ],
-        [
-            {"name": "$num_a", "description": "Number A"},
-            {"name": "$num_b"},
-        ],
-        ["$num_a", "$num_b"],
-        [
-            {"incorrect_key": "$num_a", "error": "Error"},
-        ],
-        "error",
-    ],
-)
-async def test_portia_aplan_with_plan_inputs(
-    portia: Portia,
-    planning_model: MagicMock,
-    plan_inputs: list[PlanInput] | list[dict[str, str]] | list[str],
-    telemetry: MagicMock,
-) -> None:
-    """Test async planning with various plan input formats."""
-    query = "example query"
-
-    # Mock the planning model to return a successful plan
-    planning_model.aget_structured_response.return_value = StepsOrError(steps=[], error=None)
-
-    if plan_inputs == "error":
-        with pytest.raises(ValueError, match="Invalid plan inputs received"):
-            await portia.aplan(query, plan_inputs=plan_inputs)
-    elif (
-        isinstance(plan_inputs, list)
-        and plan_inputs
-        and isinstance(plan_inputs[0], dict)
-        and "incorrect_key" in plan_inputs[0]
-    ):
-        with pytest.raises(ValueError, match="Plan input must have a name and description"):
-            await portia.aplan(query, plan_inputs=plan_inputs)
-    else:
-        plan = await portia.aplan(query, plan_inputs=plan_inputs)
-
-        telemetry.capture.assert_called_once_with(
-            PortiaFunctionCallTelemetryEvent(
-                function_name="portia_aplan",
-                function_call_details={
-                    "tools": None,
-                    "example_plans_provided": False,
-                    "end_user_provided": False,
-                    "plan_inputs_provided": True,
-                },
-                name="portia_function_call",
-            )
-        )
-
-        assert plan.plan_context.query == query
-        # Should have plan inputs
-        assert len(plan.plan_inputs) > 0
